+++ tasks/spec/fixtures/widgets/default/local_widget/require_config.js
@@ -1,0 +1,6 @@
+var require = {
+   deps: [ 'es6-shim' ],
+   paths: {
+      angular: 'new/and/improved/angular'
+   }
+};